chore: update project setup
