--- Distortion Corrector/Calib/.idea/dictionaries/liam.xml
@@ -1,8 +1,0 @@
-<component name="ProjectDictionaryState">
-  <dictionary name="liam">
-    <words>
-      <w>couldn</w>
-      <w>filepaths</w>
-    </words>
-  </dictionary>
-</component>
