--- Distortion Corrector/camera_c.json
@@ -1,1 +1,0 @@
-{"created_timestamp": "2021-03-05 11:56:58.806891", "camera_resolution": [3840, 2160], "k": [[2647.186795895339, 0.0, 1878.5274367904062], [0.0, 2669.2276077112265, 1195.2691407071964], [0.0, 0.0, 1.0]], "d": [[0.4049844460459679], [0.3045197825083126], [0.6897057892395351], [-0.39927207186605435]]}
