--- Distortion Corrector/requirements.txt
@@ -1,68 +1,0 @@
-argon2-cffi==20.1.0
-async-generator==1.10
-attrs==20.3.0
-backcall==0.2.0
-bleach==3.3.0
-certifi==2020.12.5
-cffi==1.14.5
-colorama==0.4.4
-cycler==0.10.0
-decorator==4.4.2
-defusedxml==0.7.0
-entrypoints==0.3
-ipykernel==5.5.0
-ipython==7.16.1
-ipython-genutils==0.2.0
-ipywidgets==7.6.3
-jedi==0.18.0
-Jinja2==2.11.3
-jsonschema==3.2.0
-jupyter==1.0.0
-jupyter-client==6.1.11
-jupyter-console==6.2.0
-jupyter-core==4.7.1
-jupyterlab-pygments==0.1.2
-jupyterlab-widgets==1.0.0
-kiwisolver==1.3.1
-MarkupSafe==1.1.1
-matplotlib==3.0.3
-mistune==0.8.4
-nbclient==0.5.3
-nbconvert==6.0.7
-nbformat==5.1.2
-nest-asyncio==1.5.1
-notebook==6.2.0
-nptyping==0.3.1
-numpy==1.19.5
-opencv-python==3.4.11.41
-packaging==20.9
-pandas==1.1.1
-pandocfilters==1.4.3
-parso==0.8.1
-pickleshare==0.7.5
-pigar==0.10.0
-prometheus-client==0.9.0
-prompt-toolkit==3.0.16
-pycparser==2.20
-Pygments==2.8.0
-pyparsing==2.4.7
-pyrsistent==0.17.3
-python-dateutil==2.8.1
-pytz==2021.1
-pywin32==300
-pywinpty==0.5.7
-pyzmq==22.0.3
-qtconsole==5.0.2
-QtPy==1.9.0
-scipy==1.5.2
-Send2Trash==1.5.0
-six==1.15.0
-terminado==0.9.2
-testpath==0.4.4
-tornado==6.1
-traitlets==4.3.3
-typish==1.9.1
-wcwidth==0.2.5
-webencodings==0.5.1
-widgetsnbextension==3.5.1
-wincertstore==0.2
